standalone engine encode/decode functions
